Renamed coda executable ref to mina
Went through the docs and searched for references using the coda binary
and replaced with mina. This only applies to the executable and not the
env variables or some of the tools that have not been renamed.

Additionally updated references to the client_sdk examples. Tested that
it runs properly by spinning up a node project and copying the example
code.