Fix Edit link in Mina Overview page
The URL bindings we use will not have the correct filename information
if we are on the 'Mina Overview' page. On this page, the route binding
will specify en.mdx as the file to edit which is a 404 on our Github.
Add a special check to see if we are on this page and link to the
corresponding index.mdx file on our Github.