Added support for multi language docs
This PR adds a Language context that can be set to any langauge and will
route the user to the document pages based on that selection. Currently,
we only have English docs so the routing will be based on that only.

This additionally fixes local development experience when working with
the docs since redirects were mainly used as the routing engine.