Added Language dropdown
Added a language dropdown component to allow the user to switch between
different languages. To do this, we set the language context value to be
a record that has a language and a function to change the language. We
pass this context value throughout the app to display/modify the
currently selected language.

When a user selects a new language, we compute a new URL with the old
language replaced with the newly selected language. Then we redirect the
user to that url to initiate the language change on our content.