Added a Translations solution for translations other than .mdx
Added a Translations component that handles all the non .mdx
translations in the docs website. Because we only get translations for
the .mdx content, we need another way to translate strings on the
website such as the side nav. The translations component takes a
language type and an english string and translates to the corresponding
selected language. If there are no current translations for the english
word, we return an empty string

We parse out the langauge from the url in some places due to the
Langauge context not being totally global. Due to the complexity with JS
bindings with Next, it was difficult to add this to the root level
component. We can instead just parse out the language from the url and
use that where the context cannot be used.