Refactored out local storage
Local storage was causing a bug where if a user visits the docs site and
then sets their language to russian, their language selection gets
"stuck". If their current selection is Russian but they visit the /en
route, the .mdx files will display in English but the side navs would be
russian.

To get rid of this issue, we refactor out the local storage and just use
the url parsing instead as that is much more reliable and arguably
easier to read and reason about